Add metadata_proxy_num_threads config option
This patch adds a neutron configuration option metadata_proxy_num_threads
with a default of 1000. This sets the maximum number of wsgi threads used
for the neutron metadata proxy. The neutron wide wsgi thread setting was
changed[1] to 100 which can lead to metadata proxy performance issues.
For a full discussion on the change to 100 see the dev list email[2].

[1] https://review.openstack.org/#/c/306187/
[2]http://lists.openstack.org/pipermail/openstack-dev/2015-December/082717.html