Make ns proxy use existing option for num_threads
This reverts 0e75032 in favor of using the existing option that was
intended to be used to set the wsgi pool size[1]. The problem
was neutron-ns-metadata-proxy being called without passing the
`--config-file option` so `CONF.wsgi_default_pool_size` in wsgi.Server
was always unset[2], causing it to fall back to the olso default,
even though wsgi_default_pool_size was set in neutron.conf.

Passing the config file when calling neutron-ns-metadata-proxy properly
sets the option and I have verified wsgi.Server is picking up the value
from neutron.conf.

[1] https://github.com/openstack/neutron/blob/mitaka-eol/releasenotes/notes/config-wsgi-pool-size-a4c06753b79fee6d.yaml
[2] https://github.com/openstack/neutron/blob/mitaka-eol/neutron/wsgi.py#L119